Paradigm batch: exp overflow, mid-round debuff/attack stalls, MA double-fire, message labels
Three bug reports + one label enhancement (3.88.24).

- Exp overflow (paradigm-20260916-081052): Paradigm exp passes int32 into the
  hundreds of billions, so the stat-screen parse overflowed/rejected and Exp
  drifted off gain-line accrual (Character Info showed 104M vs the real 7.05B).
  Widened PlayerStats/LastKnownStats Exp/ExpToNext/LevelExpSpan to long, the
  StatParser to long-parse (no int clamp), and the LevelUpAnnouncer/Character
  Info consumers.

- AoE debuff/attack stalls until monsters swing (paradigm-20260916-074131): the
  one-per-round between-round cast slot was gated on the client's InCombat flag,
  which flickers false during the between-kill *Combat Off*. A fresh arrival's
  pre-attack debuff re-fired into a spent round; its "already cast this round"
  rejection latched CastCoordinator's block, deferring the coupled attack to the
  next tick (the monster's swing). CastingDirector now tracks the slot time-scoped
  to the 5s round window (freed on the round tick or after the window), applying
  in or out of combat per the confirmed mechanic — no InCombat mask.

- Multi-attack fires twice on room entry (paradigm-20260916-083245): after the
  pre-attack-debuff-then-attack sequence, the attack's own *Combat Off* landed in
  the debuff's resume window and re-announced the attack. The spell resume now
  suppresses itself when the round's attack already fired after the between-round
  cast (mirrors the weapon path's ResumeAfterAttackGuard).

- Spell message "User Definitions" labels renamed for clarity: You casting /
  Casting on you / 3rd party witness / Buff/debuff applied / Buff/debuff worn off.

GAME_MECHANICS.md updated (5s between-round cadence applies in/out of combat).
5 new regression tests; dotnet build clean, 8734 tests green.

+++ Models/Profile/LastKnownStats.cs
@@ -27,11 +27,13 @@ public sealed class LastKnownStats
 
     // ----- Progression ---------------------------------------------------
     public int Level { get; set; }
-    public int Exp { get; set; }
+    // long, not int — Paradigm exp passes int32 (2.1B) into the hundreds of
+    // billions. See PlayerStats._exp.
+    public long Exp { get; set; }
     public int Lives { get; set; }
     public int Cp { get; set; }
-    public int ExpToNext { get; set; }
-    public int LevelExpSpan { get; set; }
+    public long ExpToNext { get; set; }
+    public long LevelExpSpan { get; set; }
     public int LevelPercent { get; set; }
 
     // ----- Vitals --------------------------------------------------------
